fastapi/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
boto3==1.42.1
botocore==1.42.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
Django==5.2.8
django-environ==0.12.0
django-ninja==1.5.0
ecdsa==0.19.1
fastapi==0.123.0
fonttools==4.61.0
groq==0.37.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain==1.1.2
langchain-core==1.1.1
langchain-groq==1.1.0
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.12
langsmith==0.4.55
matplotlib==3.10.7
numpy==2.3.5
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pillow==12.0.0
pyasn1==0.6.1
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
PyYAML==6.0.3
reportlab==4.4.5
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
s3transfer==0.16.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.4
starlette==0.50.0
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uuid_utils==0.12.0
uvicorn==0.38.0
xxhash==3.6.0
zstandard==0.25.0
